backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx[http2]==0.27.2
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
cryptography==43.0.1
aiofiles==24.1.0
pydantic-settings==2.5.2
